Correção de data types

+++ src/main/java/com/overdaily/Overdaily/Repository/ArmasRepository.java
@@ -3,5 +3,5 @@ package com.overdaily.Overdaily.Repository;
 import com.overdaily.Overdaily.model.Armas;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArmasRepository extends JpaRepository<Armas, Integer> {
+public interface ArmasRepository extends JpaRepository<Armas, Short> {
 }
