+++ src/main/java/com/overdaily/Overdaily/Repository/FalasRepository.java
@@ -3,5 +3,5 @@ package com.overdaily.Overdaily.Repository;
 import com.overdaily.Overdaily.model.Falas;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FalasRepository extends JpaRepository<Falas, Integer> {
+public interface FalasRepository extends JpaRepository<Falas, Short> {
 }
